Updated UseEffect call in FavoritesCard.js to avoid an infinite loop when deleting a FavroiteCard. Refactored Library.js and Favorites.js to display cards with the same spacing. Updated flex-grid spacing throughout app.
